refactor(engine): Rename engine to core
